fix: Submit 모달, Extension 모달 코드 수정
